Removes status banner from application
