+++ libs/application/templates/children-residence-change/src/lib/constants.ts
@@ -1,0 +1,10 @@
+export enum ApplicationStates {
+  DRAFT = 'draft',
+  IN_REVIEW = 'inReview',
+  SUBMITTED = 'submitted',
+}
+
+export enum Roles {
+  ParentA = 'parentA',
+  ParentB = 'parentB',
+}
